Made new simple component for UserDetails
